Add files via upload
This repository hosts my personal portfolio website, created with HTML, CSS, and JavaScript. It presents my skills, projects, and experience in a visually appealing and interactive way. The site offers an overview of my abilities, qualifications, and achievements. With a responsive design, it guarantees optimal viewing on various devices, while the combination of HTML, CSS, and JavaScript enhances user experience and interactivity.